Refactor movie actions
The movie actions was getting quite large.  So let’s break it apart!
Yay!

Pull out all the loading actions into three separate actions by type:
* movies queue
* saved movies
* dismissed movies

Yes there is some duplication of code between the files, but I think
it’s much easier to reason about in separate files completely spelled
out… and there’s not that much duplication really.  Meh.
