+++ app/actions/movie.actions.js
@@ -7,12 +7,6 @@ import {
 } from '../utils/http.utils';
 
 import {
-  LOADING_MOVIES,
-  MOVIES_QUEUE_LOADED,
-  SAVED_MOVIES_LOADED,
-  DISMISSED_MOVIES_LOADED,
-  MOVIES_ALREADY_LOADED,
-  FAILED_LOADING_MOVIES,
   SAVING_MOVIE,
   SAVED_MOVIE,
   DISMISSING_MOVIE,
@@ -22,52 +16,8 @@ import {
   FAILED_UPDATING_MOVIE,
 } from '../action-types/movie.action-types';
 
-function loadingMovies() {
-  return { type: LOADING_MOVIES };
-}
-
-function moviesAlreadyLoaded() {
-  return { type: MOVIES_ALREADY_LOADED };
-}
-
 function savingMovie() {
   return { type: SAVING_MOVIE };
-}
-
-function dismissingMovie() {
-  return { type: DISMISSING_MOVIE };
-}
-
-function undismissingMovie() {
-  return { type: UNDISMISSING_MOVIE };
-}
-
-function failedLoadingMovies(error) {
-  return {
-    type: FAILED_LOADING_MOVIES,
-    error,
-  };
-}
-
-function moviesQueueLoaded(moviesQueue) {
-  return {
-    type: MOVIES_QUEUE_LOADED,
-    moviesQueue,
-  };
-}
-
-function savedMoviesLoaded(savedMovies) {
-  return {
-    type: SAVED_MOVIES_LOADED,
-    savedMovies,
-  };
-}
-
-function dismissedMoviesLoaded(dismissedMovies) {
-  return {
-    type: DISMISSED_MOVIES_LOADED,
-    dismissedMovies,
-  };
 }
 
 function savedMovie(moviesQueue, savedMovies) {
@@ -78,6 +28,10 @@ function savedMovie(moviesQueue, savedMovies) {
   };
 }
 
+function dismissingMovie() {
+  return { type: DISMISSING_MOVIE };
+}
+
 function dismissedMovie(moviesQueue, savedMovies, dismissedMovies) {
   return {
     type: DISMISSED_MOVIE,
@@ -85,6 +39,10 @@ function dismissedMovie(moviesQueue, savedMovies, dismissedMovies) {
     savedMovies,
     dismissedMovies,
   };
+}
+
+function undismissingMovie() {
+  return { type: UNDISMISSING_MOVIE };
 }
 
 function undismissedMovie(moviesQueue, savedMovies, dismissedMovies) {
@@ -101,120 +59,6 @@ function failedUpdatingMovie(error) {
     type: FAILED_UPDATING_MOVIE,
     error,
   };
-}
-
-function fetchMoviesQueue(dispatch, getState) {
-  dispatch(loadingMovies());
-
-  const { moviesQueue, moviesQueueSkip, moviesQueueLimit } = getState().moviesState;
-
-  const uri = `/api/secure/movies?saved=false&skip=${moviesQueueSkip}` +
-    `&limit=${moviesQueueLimit}`;
-  const options = Object.assign({}, FETCH_DEFAULT_OPTIONS, {
-    method: 'GET',
-  });
-
-  return fetch(uri, options)
-    .then(checkHttpStatus)
-    .then(response => response.json())
-    .then(newMovies => {
-      const updatedMovies = unionBy(moviesQueue, newMovies, 'id');
-      return dispatch(moviesQueueLoaded(updatedMovies));
-    })
-    .catch((error) => handleHttpError(dispatch, error, failedLoadingMovies));
-}
-
-function fetchSavedMovies(dispatch, getState) {
-  dispatch(loadingMovies());
-
-  const { savedMovies, savedMoviesSkip, savedMoviesLimit } = getState().moviesState;
-
-  const uri = `/api/secure/movies?saved=true&skip=${savedMoviesSkip}` +
-    `&limit=${savedMoviesLimit}`;
-  const options = Object.assign({}, FETCH_DEFAULT_OPTIONS, {
-    method: 'GET',
-  });
-
-  return fetch(uri, options)
-    .then(checkHttpStatus)
-    .then(response => response.json())
-    .then(newMovies => {
-      const updatedMovies = unionBy(savedMovies, newMovies, 'id');
-      return dispatch(savedMoviesLoaded(updatedMovies));
-    })
-    .catch((error) => handleHttpError(dispatch, error, failedLoadingMovies));
-}
-
-function fetchDismissedMovies(dispatch, getState) {
-  dispatch(loadingMovies());
-
-  const { dismissedMovies, dismissedMoviesSkip, dismissedMoviesLimit } = getState().moviesState;
-
-  const uri = `/api/secure/movies?dismissed=true&skip=${dismissedMoviesSkip}` +
-    `&limit=${dismissedMoviesLimit}`;
-  const options = Object.assign({}, FETCH_DEFAULT_OPTIONS, {
-    method: 'GET',
-  });
-
-  return fetch(uri, options)
-    .then(checkHttpStatus)
-    .then(response => response.json())
-    .then(newMovies => {
-      const updatedMovies = unionBy(dismissedMovies, newMovies, 'id');
-      return dispatch(dismissedMoviesLoaded(updatedMovies));
-    })
-    .catch((error) => handleHttpError(dispatch, error, failedLoadingMovies));
-}
-
-export function loadInitialMoviesQueue() {
-  return (dispatch, getState) => {
-    const { moviesQueue } = getState().moviesState;
-
-    // check to see if we have any movies, and if so, stop here
-    if (moviesQueue && moviesQueue.length > 0) {
-      return dispatch(moviesAlreadyLoaded());
-    }
-
-    return fetchMoviesQueue(dispatch, getState);
-  };
-}
-
-export function loadMoviesQueue() {
-  return fetchMoviesQueue;
-}
-
-export function loadInitialSavedMovies() {
-  return (dispatch, getState) => {
-    const { savedMovies } = getState().moviesState;
-
-    // check to see if we have any movies, and if so, stop here
-    if (savedMovies && savedMovies.length > 0) {
-      return dispatch(moviesAlreadyLoaded());
-    }
-
-    return fetchSavedMovies(dispatch, getState);
-  };
-}
-
-export function loadSavedMovies() {
-  return fetchSavedMovies;
-}
-
-export function loadInitialDismissedMovies() {
-  return (dispatch, getState) => {
-    const { dismissedMovies } = getState().moviesState;
-
-    // check to see if we have any movies, and if so, stop here
-    if (dismissedMovies && dismissedMovies.length > 0) {
-      return dispatch(moviesAlreadyLoaded());
-    }
-
-    return fetchDismissedMovies(dispatch, getState);
-  };
-}
-
-export function loadDismissedMovies() {
-  return fetchDismissedMovies;
 }
 
 export function saveMovie(movieId) {
